Match INT_MAX check in mbed_minimal_formatted_string with calling code
Otherwise we get no output for functions which print to stdout or a buffer of
unrestricted length on platforms where LONG_MAX > INT_MAX.

This is e.g. the case when compiling for x64 Linux and any other platform where
sizeof(long) > sizeof(int).

+++ platform/source/minimal-printf/mbed_printf_wrapper.c
@@ -34,7 +34,7 @@ int PREFIX(printf)(const char *format, ...)
 {
     va_list arguments;
     va_start(arguments, format);
-    int result = mbed_minimal_formatted_string(NULL, LONG_MAX, format, arguments, stdout);
+    int result = mbed_minimal_formatted_string(NULL, INT_MAX, format, arguments, stdout);
     va_end(arguments);
 
     return result;
@@ -44,7 +44,7 @@ int PREFIX(sprintf)(char *buffer, const char *format, ...)
 {
     va_list arguments;
     va_start(arguments, format);
-    int result = mbed_minimal_formatted_string(buffer, LONG_MAX, format, arguments, NULL);
+    int result = mbed_minimal_formatted_string(buffer, INT_MAX, format, arguments, NULL);
     va_end(arguments);
 
     return result;
@@ -62,12 +62,12 @@ int PREFIX(snprintf)(char *buffer, size_t length, const char *format, ...)
 
 int PREFIX(vprintf)(const char *format, va_list arguments)
 {
-    return mbed_minimal_formatted_string(NULL, LONG_MAX, format, arguments, stdout);
+    return mbed_minimal_formatted_string(NULL, INT_MAX, format, arguments, stdout);
 }
 
 int PREFIX(vsprintf)(char *buffer, const char *format, va_list arguments)
 {
-    return mbed_minimal_formatted_string(buffer, LONG_MAX, format, arguments, NULL);
+    return mbed_minimal_formatted_string(buffer, INT_MAX, format, arguments, NULL);
 }
 
 int PREFIX(vsnprintf)(char *buffer, size_t length, const char *format, va_list arguments)
@@ -79,7 +79,7 @@ int PREFIX(fprintf)(FILE *stream, const char *format, ...)
 {
     va_list arguments;
     va_start(arguments, format);
-    int result = mbed_minimal_formatted_string(NULL, LONG_MAX, format, arguments, stream);
+    int result = mbed_minimal_formatted_string(NULL, INT_MAX, format, arguments, stream);
     va_end(arguments);
 
     return result;
@@ -87,7 +87,7 @@ int PREFIX(fprintf)(FILE *stream, const char *format, ...)
 
 int PREFIX(vfprintf)(FILE *stream, const char *format, va_list arguments)
 {
-    return mbed_minimal_formatted_string(NULL, LONG_MAX, format, arguments, stream);
+    return mbed_minimal_formatted_string(NULL, INT_MAX, format, arguments, stream);
 }
 
 #endif // MBED_MINIMAL_PRINTF
